+++ flutter_map/lib/src/core/center_zoom.dart
@@ -1,0 +1,7 @@
+import 'package:latlong/latlong.dart';
+
+class CenterZoom {
+  final LatLng center;
+  final double zoom;
+  CenterZoom({this.center, this.zoom});
+}