+++ flutter_map/lib/src/map/map.dart
@@ -1,19 +1,43 @@
 import 'dart:async';
+import 'dart:math' as math;
 
+import 'package:flutter_map/src/core/center_zoom.dart';
 import 'package:latlong/latlong.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map/src/core/bounds.dart';
 import 'package:flutter_map/src/core/point.dart';
 
 class MapControllerImpl implements MapController {
-  MapState state;
+  Completer<Null> _readyCompleter = new Completer<Null>();
+  MapState _state;
+
+  Future<Null> get onReady => _readyCompleter.future;
+  set state(MapState state) {
+    _state = state;
+    if (!_readyCompleter.isCompleted) {
+      _readyCompleter.complete();
+    }
+  }
 
   void move(LatLng center, double zoom) {
-    state.move(center, zoom);
+    _state.move(center, zoom);
   }
+
+  void fitBounds(
+    LatLngBounds bounds, {
+    FitBoundsOptions options =
+        const FitBoundsOptions(padding: const Point(24.0, 24.0)),
+  }) {
+    _state.fitBounds(bounds, options);
+  }
+
+  bool get ready => _state != null;
+
+  LatLng get center => _state.center;
+  double get zoom => _state.zoom;
 }
 
-class MapState  {
+class MapState {
   final MapOptions options;
   final StreamController<Null> _onMoveSink;
 
@@ -60,6 +84,18 @@ class MapState  {
     _lastCenter = center;
     _pixelOrigin = getNewPixelOrigin(center);
     _onMoveSink.add(null);
+
+    if (options.onPositionChanged != null) {
+      options.onPositionChanged(new MapPosition(center: center, zoom: zoom));
+    }
+  }
+
+  void fitBounds(LatLngBounds bounds, FitBoundsOptions options) {
+    if (!bounds.isValid) {
+      throw ("Bounds are not valid.");
+    }
+    var target = _getBoundsCenterZoom(bounds, options);
+    move(target.center, target.zoom);
   }
 
   LatLng getCenter() {
@@ -67,6 +103,42 @@ class MapState  {
       return _lastCenter;
     }
     return layerPointToLatLng(_centerLayerPoint);
+  }
+
+  CenterZoom _getBoundsCenterZoom(
+      LatLngBounds bounds, FitBoundsOptions options) {
+    var paddingTL = options.padding;
+    var paddingBR = options.padding;
+
+    var zoom = getBoundsZoom(bounds, paddingTL + paddingBR, inside: false);
+    zoom = math.min(options.maxZoom, zoom);
+
+    var paddingOffset = (paddingBR - paddingTL) / 2;
+    var swPoint = project(bounds.southWest, zoom);
+    var nePoint = project(bounds.northEast, zoom);
+    var center = unproject((swPoint + nePoint) / 2 + paddingOffset, zoom);
+    return new CenterZoom(
+      center: center,
+      zoom: zoom,
+    );
+  }
+
+  double getBoundsZoom(LatLngBounds bounds, Point<double> padding,
+      {bool inside = false}) {
+    var zoom = this.zoom ?? 0.0;
+    var min = this.options.minZoom ?? 0.0;
+    var max = this.options.maxZoom ?? double.infinity;
+    var nw = bounds.northWest;
+    var se = bounds.southEast;
+    var size = this.size - padding;
+    var boundsSize = new Bounds(project(se, zoom), project(nw, zoom)).size;
+    var scaleX = size.x / boundsSize.x;
+    var scaleY = size.y / boundsSize.y;
+    var scale = inside ? math.max(scaleX, scaleY) : math.min(scaleX, scaleY);
+
+    zoom = getScaleZoom(scale, zoom);
+
+    return math.max(min, math.min(max, zoom));
   }
 
   Point project(LatLng latlng, [double zoom]) {
@@ -95,6 +167,12 @@ class MapState  {
     var crs = options.crs;
     fromZoom = fromZoom == null ? _zoom : fromZoom;
     return crs.scale(toZoom) / crs.scale(fromZoom);
+  }
+
+  double getScaleZoom(double scale, double fromZoom) {
+    var crs = options.crs;
+    fromZoom = fromZoom == null ? _zoom : fromZoom;
+    return crs.zoom(scale * crs.scale(fromZoom));
   }
 
   Bounds getPixelWorldBounds(double zoom) {