+++ flutter_map/CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.0.5] - 3/11/2018
+- fitBounds, onPositionChanged (#39)
+
+## [0.0.5] - 3/11/2018
+
+- make tile background customizable (#36)
+- use transparent_image as placeholder image (#37)
+
 ## [0.0.4] - 4/18/2018
 
 - Add marker anchor support (#27, #30)